requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sendgrid>=6.11.0
openai>=1.0.0
